Refactor and fix LightManager synchronization issues
- Move inner classes in LightManager to outer scope.
- Deprecate old light type classes.
- Reload light configuration file in the update function to synchronize
  with the client's render loop.
