+++ src/main/java/rs117/hd/lighting/LightType.java
@@ -1,0 +1,6 @@
+package rs117.hd.lighting;
+
+public enum LightType
+{
+	STATIC, FLICKER, PULSE
+}
